sound pool and some more optimizations

+++ src/ecs/ecs_component_сontainer_index.tpp
@@ -10,7 +10,7 @@ namespace omniscia::core::ecs {
     template<typename T> 
     class ECS_Index {
         private:
-            u32 _index;
+            u8 _index;
 
             bool _success;
         public:
